Scheduler configurations moved to properties file
