+++ src/main/resources/application-dev.properties
@@ -55,6 +55,10 @@ mediflow.billing.amount=150
 mediflow.pdf.path = C:/Logs/Invoices
 mediflow.sensitive.backup.path = C:/Logs/SensitiveData
 
+mediflow.scheduler.invoice.pdf=0 0 2 * * *
+mediflow.scheduler.overdue.payment=0 0 2 * * *
+mediflow.scheduler.clean.backup=0 0 2 * * *
+
 ## -----------------------------
 ## DETAILED EXCEPTION PROPERTIES
 ## -----------------------------
